+++ backend/prisma/migrations/20260421054852_add_tutor_profile/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "users" ADD COLUMN     "subjects" TEXT[],
+ADD COLUMN     "teachGrades" INTEGER[];